+++ src/Microsoft.Android.Build.BaseTasks/global.json
